+++ R/bikm1-package.R
@@ -29,7 +29,7 @@
 ##'
 ##' Robert. Classification croisee pour l'analyse de bases de donnees de grandes dimensions de pharmacovigilance. Thesis, Paris Saclay (2017).
 ##'
-##' Robert and Vasseur. Comparing high dimensional partitions with the Co-clustering Adjusted Rand Index, Preprint (2017).
+##' Robert, Vasseur and Brault. Comparing high dimensional partitions with the Co-clustering Adjusted Rand Index, Journal of Classification, 38(1), 158-186 (2021).
 ##'
 ##' @import ade4
 ##' @import pracma
@@ -37,6 +37,7 @@
 ##' @import ggplot2
 ##' @import reshape2
 ##' @import grid
+##' @importFrom lpSolve lp.assign
 ##' @importFrom parallel mclapply detectCores
 ##' @importFrom grDevices dev.new
 ##' @importFrom graphics boxplot lines par title mtext