+++ NEWS.md
@@ -4,5 +4,5 @@
 
 Changes:
 
-* Modified examples to reduce running time and fix misspelled variables
-* Bugfixes in the plot functions
+* New index fonctions added: MI_simple, ENMI, CoNMI, NCE_simple, NCE_LBM. For more details see: Comparing high-dimensional partitions with the Co-clustering Adjusted Rand Index. Robert V., Vasseur Y. and Brault V., Journal of classification 38 (1), 158-186 (2021).
+*